created initial datamodel for unified data layer
This commit focuses on creating an initial datamodel for the unified
data layer. This focuses on the types and not:
  * mutations
  * filters
  * subscriptions
Once we can agree on the types as a second step we can create the
mutations, filters, and subscriptions needed.

I also removed some of the old OLI code in this commit as this is
refactoring that API.

Issue: NH-132
Test Plan: Running build containers, push helm, and cicd. Also viewed
  schema is voyager.